requirements.txt
streamlit
pytesseract
opencv-python
Pillow
thefuzz
pandas
python-Levenshtein
